+++ packages/create/src/new-consumer-project.ts
@@ -1,0 +1,149 @@
+/**
+ * @license
+ * Copyright 2022-2024 Matter.js Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { execSync } from "child_process";
+import { cp, writeFile } from "fs/promises";
+import { basename, resolve } from "path";
+import { Config, Template, TEMPLATE_DIR } from "./config.js";
+import { build, createAndValidateDest, createVsCodeProject, TemplateNotFoundError } from "./new-project.js";
+
+const PACKAGE_JSON = {
+    dependencies: {} as Record<string, string>,
+    devDependencies: {
+        typescript: "*",
+        "@types/node": "*",
+    },
+    name: "matter-app",
+    version: "0.1.0",
+    type: "module",
+    scripts: {
+        prepare: "tsc",
+        build: "tsc",
+        app: "node --enable-source-maps dist/index.js",
+        test: 'echo "Error: no test specified" && exit 1',
+        clean: "node -e 'fs.rmSync(\"dist\", { recursive: true, force: true })'",
+    },
+};
+
+const TSCONFIG = {
+    compilerOptions: {
+        strict: true,
+        target: "es2022",
+        module: "node16",
+        moduleResolution: "node16",
+        noImplicitAny: true,
+        noImplicitOverride: true,
+        outDir: "./dist",
+        skipLibCheck: true,
+    },
+    include: ["./src/**/*.ts"],
+};
+
+const GITIGNORE = "node_modules/\ndist/\n";
+
+export interface NewConsumerProject {
+    kind: "consumer";
+    template: Template;
+    source: string;
+    dest: string;
+
+    create(): Promise<void>;
+    build(showInstall: boolean): void;
+}
+
+export async function NewConsumerProject(name: string, dest: string): Promise<NewConsumerProject> {
+    const template = (await Config()).templates.find(template => template.name === name);
+
+    if (template === undefined) {
+        throw new TemplateNotFoundError(`"${name}" is not a valid template name`);
+    }
+
+    return {
+        kind: "consumer",
+        template,
+        source: resolve(TEMPLATE_DIR, template.name),
+        dest,
+        create: create,
+        build,
+    };
+}
+
+export async function create(this: NewConsumerProject) {
+    await createAndValidateDest(this);
+    await createPackageJson(this);
+    await createTsconfig(this);
+    await createGitignore(this);
+    await installSources(this);
+    await createConsumerVsCodeProject(this);
+}
+
+async function createPackageJson(project: NewConsumerProject) {
+    const pkg = PACKAGE_JSON;
+
+    const config = await Config();
+
+    pkg.scripts.app = `node --enable-source-maps ${entrypointFor(project)}`;
+
+    pkg.dependencies = project.template.dependencies;
+    pkg.devDependencies["typescript"] = config.typescriptVersion;
+    pkg.devDependencies["@types/node"] = config.nodeTypesVersion;
+
+    (pkg as any).description = project.template.description;
+
+    let author, authorEmail;
+    try {
+        for (const line of execSync("git config --list", { encoding: "utf-8" }).split("\n")) {
+            if (line.startsWith("user.name=")) {
+                author = line.replace(/^[^=]=/, "");
+            } else if (line.startsWith("user.email=")) {
+                authorEmail = line.replace(/^[^=]=/, "");
+            }
+        }
+    } catch (e) {
+        // Non-fatal error if git isn't available; we just don't set the author
+    }
+
+    if (authorEmail) {
+        if (author) {
+            author = `${author} <${authorEmail}>`;
+        } else {
+            author = `<${authorEmail}>`;
+        }
+    }
+
+    if (author) {
+        (pkg as any).author = author;
+    }
+
+    await writeFile(resolve(project.dest, "package.json"), JSON.stringify(pkg, undefined, 4));
+}
+
+async function createTsconfig(project: NewConsumerProject) {
+    await writeFile(resolve(project.dest, "tsconfig.json"), JSON.stringify(TSCONFIG, undefined, 4));
+}
+
+async function createGitignore(project: NewConsumerProject) {
+    await writeFile(resolve(project.dest, ".gitignore"), GITIGNORE);
+}
+
+async function installSources(project: NewConsumerProject) {
+    await cp(project.source, resolve(project.dest, "src"), {
+        recursive: true,
+        filter: source => basename(source).indexOf(".") === -1 || source.endsWith(".ts"),
+    });
+}
+
+async function createConsumerVsCodeProject(project: NewConsumerProject) {
+    async function* launchGenerator() {
+        yield { name: "Run Program", program: entrypointFor(project) };
+    }
+
+    await createVsCodeProject(project, launchGenerator);
+}
+
+function entrypointFor(project: NewConsumerProject) {
+    return "dist/" + project.template.entrypoint.replace(/\.ts$/, ".js");
+}